% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alljuvtt.R
\docType{data}
\name{alljuvtt}
\alias{alljuvtt}
\title{alljuvtt}
\format{a dataframe with 3812 observations of 10 variables}
\source{
alljuvtt <- readRDS('~/Dropbox/GitHubRepos/JuvSalmon_Manuscript/data/allfishtraveltime_starterdf.rds')
}
\usage{
alljuvtt
}
\description{
Dataframe of fishpaths for the juvenile salmon data (both Sac Bank fish and Yolo fish)
}
\keyword{datasets}
