% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputData.R
\name{checkInputData}
\alias{checkInputData}
\title{checkInputData}
\usage{
checkInputData(dTagID, dDets, dStation, dRkm)
}
\arguments{
\item{dTagID}{TagID column of your starting dataframe of TagIDs, detections, and river kilometers.}
}
\value{
nothing, if it's all good. If not, throws an error for the first column that's not good.
}
\description{
checkInputData
}
\details{
function author: Myfanwy Johnston
}
