% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFirstLast_internal.R
\name{firstlastOneFish}
\alias{firstlastOneFish}
\title{getFirstLast_internal; called by getFirstLast}
\usage{
firstlastOneFish(x, tagidcol, datetimecol = dtc2, stationcol = stnc2)
}
\description{
getFirstLast_internal; called by getFirstLast
}
