% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCDEC.R
\name{get.CDEC}
\alias{get.CDEC}
\title{get.CDEC}
\usage{
get.CDEC(station, duration, sensor, start, end)
}
\arguments{
\item{station}{the number of the station you want to query (ex: "LIS")}

\item{duration}{the type of recording that station does (example is "E" for event, "D" for daily, etc)}

\item{sensor}{the number of the sensor you want, see http://cdec.water.ca.gov/misc/senslist.html}

\item{start}{start date in "yyyy-mm-dd" format}

\item{end}{end date in "yyyy-mm-dd" format}
}
\value{
asks if you want to write to a csv
}
\description{
get.CDEC
}
\examples{
# test with Lisbon Weir:
get.CDEC("LIS","E",61,"2014-09-30","2014-11-21")
}

