% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFirstLast.R
\name{getFirstLast}
\alias{getFirstLast}
\title{getFirstLast}
\usage{
getFirstLast(df)
}
\arguments{
\item{df}{the output of the fishpaths() function: a dataframe that includes of TagIDs, arrival, departure, TagID, and Rkm.}
}
\value{
df with TagID, station of first detection, station of last detection, travel time in days between those two stations, and the distance between the two stations.
}
\description{
getFirstLast
}
\details{
function author: Myfanwy Johnston
}
