% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReachMovements.R
\name{getReachOneFish}
\alias{getReachOneFish}
\title{getReachMovements}
\usage{
getReachOneFish(x)
}
\arguments{
\item{x}{a dataframe that is the output of the fishpaths() function}
}
\value{
a dataframe with the TagID, departure station, arrival station, and travel time in days
}
\description{
getReachMovements
}
\details{
function author: Matt Espe
}
