% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReachTravelTime.R
\name{getReachTravelTime}
\alias{getReachTravelTime}
\title{getReachTravelTime}
\usage{
getReachTravelTime(df, s1, s2)
}
\arguments{
\item{df}{dataframe of detections}

\item{s1}{station at the top of the reach, in quotes (must match the name of the station in df)}

\item{s2}{station at the bottom of the reach, in quotes (must match the name of the station in df)}
}
\value{
a dataframe with the TagID, travel time in days (column "tt"), Station1, Station2, and a generated "reachname" column.
}
\description{
A function to calculate the time (in days) between detections at the top and bottom of a reach.
}
