% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishpaths.R
\name{fishpaths}
\alias{fishpaths}
\title{fishpaths}
\usage{
fishpaths(x, TagID_col, Station_col, Datetime_col)
}
\arguments{
\item{x}{a dataframe of detections}

\item{TagID_col}{must be called "TagID"}

\item{Station_col}{must be called "Station"}

\item{Datetime_col}{indicate the column that contains detection date and time; must be in POSIXct format (YYYY-MM-DD HH:MM:SS or similar)}
}
\value{
dataframe with fishpaths for each tagID
}
\description{
fishpaths
}
\examples{
# Test
library(ybp)
f <- all69khz_grouped
fishpaths(f, f$TagID, f$Station)
}
