% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishpaths.R
\name{fishpaths}
\alias{fishpaths}
\title{fishpaths}
\usage{
fishpaths(x, TagID_col, Station_col)
}
\arguments{
\item{x}{a dataframe of detections}

\item{TagID_col}{must be called "TagID"}

\item{Station_col}{must be called "Station"}
}
\value{
dataframe with fishpaths for each tagID
}
\description{
fishpaths
}
\details{
For current version of function, rest of the columns must be present in the dataframe with the following names: "DateTimeUTC", "Receiver",  "TagID",   "Station" , "Rkm" ,   "lat" ,    "lon",   "Sp", "DateTagged".  This will be simplified to just "DateTimeUTC", "TagID" and "Station" soon.
}
\examples{
# Test
library(ybp)
f <- all69khz_grouped
fishpaths(f, f$TagID, f$Station)
}

