% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishpaths.R
\name{fishpaths}
\alias{fishpaths}
\title{fishpaths}
\usage{
fishpaths(x, TagID_col, Station_col)
}
\arguments{
\item{Station_col}{}
}
\value{
dataframe with fishpaths for each tagID
}
\description{
fishpaths
}
\examples{
# Test
library(ybp)
f <- all69khz_grouped
fishpaths(f, f$TagID, f$Station)
}

