% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFirstLast.R
\name{getFirstLast}
\alias{getFirstLast}
\title{getFirstLast}
\usage{
getFirstLast(df, tidc2 = "TagID", dtc2 = "DateTimeUTC",
  stnc2 = "Station")
}
\arguments{
\item{df}{the output of the fishpaths() function: a dataframe that includes of TagIDs, arrival, departure, TagID, and Rkm.}
}
\value{
df with TagID, station of first detection, station of last detection, travel time in days between those two stations, and the distance between the two stations.
}
\description{
getFirstLast
}
\details{
Function selects the first station as that with the earliest time, breaking ties with the lowest river km, and the last station as that with the last time, with the highest river km.
}
\author{
Myfanwy Johnston
}
